feat: avoid flagging comments that contain parsable code
In several scenarios developers needs to be able to comment out code,
either temporarily or permanently. Before this commit the plugin
would attempt to fix these snippets, but this is not desirable since
it potentially can alter the semantics of the commented program.

Now we check the comment if it contains parsable code, and if it does
we treat it as an unfixable comment